+++ infrastructure/src/main/java/co/com/soft/repository/UserRepository.java
@@ -3,5 +3,5 @@ package co.com.soft.repository;
 import co.com.soft.entity.UserEntity;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface UserRepository extends ReactiveCrudRepository<UserEntity, Integer> {
+public interface UserRepository extends ReactiveCrudRepository<UserEntity, Long> {
 }
